+++ src/commands/export.ts
@@ -64,7 +64,16 @@ const chromiumRequiredExtensions = [
   ...extensions.jpeg,
 ] as string[]
 
-export const doExport = async (uri: Uri, document: TextDocument) => {
+export interface ExportResult {
+  uri: Uri
+  autoOpen?: boolean
+  error?: string
+}
+
+export const doExport = async (
+  uri: Uri,
+  document: TextDocument,
+): Promise<ExportResult> => {
   let proxyServer: WorkspaceProxyServer | undefined
   let baseUrl: string | undefined
 
@@ -148,7 +157,7 @@ export const doExport = async (uri: Uri, document: TextDocument) => {
         pptxEditable,
       }: {
         pptxEditable?: boolean
-      }) => {
+      }): Promise<ExportResult> => {
         const conf = await createConfigFile(document, {
           allowLocalFiles: !proxyServer,
           pdfNotes:
@@ -237,31 +246,28 @@ export const doExport = async (uri: Uri, document: TextDocument) => {
           }
         }
 
-        if (
-          marpConfiguration().get<boolean>('exportAutoOpen') &&
-          outputToLocalFS // Only local files can open in a relevant application
-        ) {
-          env.openExternal(uri)
-        } else {
-          // Show success message if the auto open preference is disabled, or
-          // the output is not local file: Remote path, Virtual file system, etc.
-          window.showInformationMessage(
-            `Marp slide deck was successfully exported to ${uri.toString()}.`,
-          )
+        return {
+          uri,
+          autoOpen:
+            marpConfiguration().get<boolean>('exportAutoOpen') &&
+            outputToLocalFS, // Only local files can open in a relevant application
         }
       }
 
-      await runMarpCli({ pptxEditable: pptxEditableSmart ? true : undefined })
+      return await runMarpCli({
+        pptxEditable: pptxEditableSmart ? true : undefined,
+      })
     } catch (e) {
-      window.showErrorMessage(
-        `Failure to export${(() => {
+      return {
+        uri,
+        error: `Failure to export${(() => {
           if (e instanceof MarpCLIError) return `. ${e.message}`
           if (e instanceof Error) return `: [${e.name}] ${e.message}`
           if (hasToString(e)) return `. ${e.toString()}`
 
           return ' by unknown error.'
         })()}`,
-      )
+      }
     } finally {
       input.cleanup()
     }
@@ -293,13 +299,25 @@ export const saveDialog = async (document: TextDocument) => {
   })
 
   if (saveURI) {
-    await window.withProgress(
+    const result = await window.withProgress(
       {
         location: ProgressLocation.Notification,
         title: `Exporting Marp slide deck to ${saveURI.toString()}...`,
       },
-      () => doExport(saveURI, document),
+      async () => await doExport(saveURI, document),
     )
+
+    if (result.error) {
+      window.showErrorMessage(result.error)
+    } else if (result.autoOpen) {
+      env.openExternal(result.uri)
+    } else {
+      // Show success message if the auto open was disabled. It includes
+      // when the output is not local file: Remote path, Virtual file, etc.
+      window.showInformationMessage(
+        `Marp slide deck was successfully exported to ${result.uri.toString()}.`,
+      )
+    }
   }
 }